+++ backend/libs/todo/todo-persistence/src/lib/schema/todo.schema.ts
@@ -14,9 +14,6 @@ class TodoItem {
   done?: boolean;
 
   @Prop({ required: false })
-  amount?: string;
-
-  @Prop({ required: false })
   category?: string;
 }